Rename all .jsx files to .js, update all requires
