+++ app/Grid/Components.js
@@ -4,14 +4,14 @@
  * @memberOf module:Grid
  */
 var Components = {
-    Cell: require('./Components/Cell.jsx'),
-    MainGrid: require('./Components/MainGrid.jsx'),
-    Mixins: require('./Components/Mixins.js'),
-    ModulesCache: require('./Components/ModulesCache.js'),
-    ModuleHolder: require('./Components/ModuleHolder.jsx'),
-    Resizer: require('./Components/Resizer.jsx'),
-    Row: require('./Components/Row.jsx'),
-    SubGrid: require('./Components/SubGrid.jsx'),
+    Cell: require('./Components/Cell'),
+    MainGrid: require('./Components/MainGrid'),
+    Mixins: require('./Components/Mixins'),
+    ModulesCache: require('./Components/ModulesCache'),
+    ModuleHolder: require('./Components/ModuleHolder'),
+    Resizer: require('./Components/Resizer'),
+    Row: require('./Components/Row'),
+    SubGrid: require('./Components/SubGrid'),
 };
 
 module.exports = Components;
