+++ app/Grid/Modules.js
@@ -4,8 +4,8 @@
  * @memberOf module:Grid
  */
 var Modules = {
-    Test1: require('./Modules/Test1.jsx'),
-    Test2: require('./Modules/Test2.jsx'),
+    Test1: require('./Modules/Test1'),
+    Test2: require('./Modules/Test2'),
 };
 
 module.exports = Modules;
