+++ app/Grid/Modules/Test2.js
@@ -1,18 +1,18 @@
 /** @jsx React.DOM */
 var React = require('react/addons');  // react + addons
 
-var ModuleMixin = require('./Mixins/Module.jsx');
+var ModuleMixin = require('./Mixins/Module');
 
 
-var Test1 = {
+var Test2 = {
     mixins: [
         ModuleMixin,
     ],
 
     renderModule: function() {
-        return 'Module.Test1: ' + this.props.text;
+        return 'Module.Test2: ' + this.props.text;
     }
 };
 
-module.exports = Test1 = React.createClass(Test1);
+module.exports = Test2 = React.createClass(Test2);
