+++ app/main.js
@@ -1,6 +1,6 @@
 /** @jsx React.DOM */
 var React = require('react/addons');  // react + addons
 
-var App = require('./App.jsx');
+var App = require('./App');
 
 React.render(<App/>, document.getElementById('app'));
