+++ specs/Grid/Components/ModuleHolder-spec.js
@@ -4,15 +4,15 @@ var React = require('react/addons');  // react + addons
 var stringify = require('json-stable-stringify');
 var TestUtils = React.addons.TestUtils;
 
-var Manipulator = require('../../../app/Grid/Manipulator.js');
-var Modules = require('../../../app/Grid/Modules.js');
-var Store = require('../../../app/Grid/Store.js');
+var Manipulator = require('../../../app/Grid/Manipulator');
+var Modules = require('../../../app/Grid/Modules');
+var Store = require('../../../app/Grid/Store');
 
-var ModulesCache = require('../../../app/Grid/Components/ModulesCache.js');
-var ModuleHolder = require('../../../app/Grid/Components/ModuleHolder.jsx');
+var ModulesCache = require('../../../app/Grid/Components/ModulesCache');
+var ModuleHolder = require('../../../app/Grid/Components/ModuleHolder');
 
-var Utils = require('../../Utils.js');
-var componentUtils = require('./Utils.js');
+var Utils = require('../../Utils');
+var componentUtils = require('./Utils');
 
 
 describe("Grid.Components.ModuleHolder", function() {
